Refactor SaveDir in archiver
Move SaveDir back to archiver and use hooks in the os specific archivers. Rename _unix to _notwin.
