+++ internal/archiver/archiver_windows.go
@@ -1,15 +1,12 @@
 package archiver
 
 import (
-	"context"
 	"os"
 	"path/filepath"
-	"sort"
 
 	"github.com/restic/restic/internal/debug"
 	"github.com/restic/restic/internal/errors"
 	"github.com/restic/restic/internal/fs"
-	"github.com/restic/restic/internal/restic"
 )
 
 // preProcessTargets performs preprocessing of the targets before the loop.
@@ -33,66 +30,28 @@ func processTarget(_ fs.FS, target string) string {
 	return target
 }
 
-// SaveDir stores a directory in the repo and returns the node. snPath is the
-// path within the current snapshot. In case of windows it also adds the ads
-// files as top level nodes. We selectively filter out the ads files later
-// for functionalities like totalCount, fileCount, filtering etc.
-func (arch *Archiver) SaveDir(ctx context.Context, snPath string, dir string, fi os.FileInfo, previous *restic.Tree, complete CompleteFunc) (d FutureNode, err error) {
-	debug.Log("%v %v", snPath, dir)
+// getNameFromPathname gets the name from pathname.
+// In case for windows the pathname is the full path, so it need to get the base name.
+func getNameFromPathname(pathname string) (name string) {
+	return filepath.Base(pathname)
+}
 
-	treeNode, err := arch.nodeFromFileInfo(snPath, dir, fi)
-	if err != nil {
-		return FutureNode{}, err
-	}
+// preProcessPaths processes paths before looping.
+func (arch *Archiver) preProcessPaths(dir string, names []string) (paths []string) {
+	// In case of windows we want to add the ADS paths as well before sorting.
+	return arch.getPathsIncludingADS(dir, names)
+}
 
-	names, err := readdirnames(arch.FS, dir, fs.O_NOFOLLOW)
-	if err != nil {
-		return FutureNode{}, err
-	}
-	//In case of windows we want to add the ADS paths as well before sorting.
-	paths := getPathsIncludingADS(arch, dir, names)
-	sort.Strings(paths)
-
-	nodes := make([]FutureNode, 0, len(paths))
-
-	for _, pathname := range paths {
-		// test if context has been cancelled
-		if ctx.Err() != nil {
-			debug.Log("context has been cancelled, aborting")
-			return FutureNode{}, ctx.Err()
-		}
-
-		name := filepath.Base(pathname)
-		oldNode := previous.Find(name)
-		snItem := join(snPath, name)
-		fn, excluded, err := arch.Save(ctx, snItem, pathname, oldNode)
-
-		// return error early if possible
-		if err != nil {
-			err = arch.error(pathname, err)
-			if err == nil {
-				// ignore error
-				continue
-			}
-
-			return FutureNode{}, errors.Wrap(err, "error saving a target (file or directory)")
-		}
-
-		if excluded {
-			continue
-		}
-
-		nodes = append(nodes, fn)
-	}
-
-	fn := arch.treeSaver.Save(ctx, snPath, dir, treeNode, nodes, complete)
-
-	return fn, nil
+// processPath processes the path in the loop.
+func (arch *Archiver) processPath(_ string, name string) (path string) {
+	// In case of windows we have already prepared the paths before the loop.
+	// Hence this is a no-op.
+	return name
 }
 
 // getPathsIncludingADS iterates all passed path names and adds the ads
 // contained in those paths before returning all full paths including ads
-func getPathsIncludingADS(arch *Archiver, dir string, names []string) []string {
+func (arch *Archiver) getPathsIncludingADS(dir string, names []string) []string {
 	paths := make([]string, 0, len(names))
 
 	for _, name := range names {
